requirements.txt
scikit-learn==1.0.2
pandas==1.3.5
numpy==1.21.6
networkx==2.6.3
scipy==1.7.3
torch==1.12.1+cu113
dgl-cu113==0.8.1